Update with upstream master.

+++ SYSTEM_ADMIN/system_info/data_source/data_source.c
@@ -1,6 +1,7 @@
 #include "data_source.h"
 #include "system_log.h"
 #include <mysql/mysql.h>
+#include <string.h>
 
 #define RECORD_STATUS_NORMAL                0
 
@@ -79,7 +80,7 @@ BOOL data_source_system_info(int *preal_domains, int *pbackup, int *pmonitor,
 	int *punchkusr, int *psubsys, int *psms, int *pextpasswd, int *palias_domains,
 	int *poutofdate, int *pdeleted, int *psuspend, int *pgroups,
 	int *palloc_addresses, int *preal_addresses, int *palias_address,
-	int *pmlists, int *ptotal_space)
+	int *pmlists, long *ptotal_space)
 {
 	int i, j, rows;
 	int temp_type;
